comments

+++ makefile
@@ -7,5 +7,6 @@ typemanager:
 	ghc main.hs -o TypeManager
 
 clean:
-	rm TypeManager Tests *.o *.hi *.tix 
-	rm -r .hpc
+	rm TypeManager Tests *.o *.hi 
+	if test -s "Tests.tix"; then rm Tests.tix; fi
+	if test -s ".hpc"; then rm -r .hpc; fi
